add scroll handler

+++ includes/ft_typedefs.h
@@ -6,7 +6,7 @@
 /*   By: flfische <flfische@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 09:32:00 by flfische          #+#    #+#             */
-/*   Updated: 2024/04/17 09:32:48 by flfische         ###   ########.fr       */
+/*   Updated: 2024/04/17 10:11:20 by flfische         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ typedef struct s_fractol
 	int			max_iter;
 	uint32_t	(*draw_fractal)(struct s_fractol *fractol, int x, int y);
 	t_mouse		mouse;
-	float		zoom;
+	double		zoom;
 	t_complex	offset;
 }				t_fractol;
